New: clearBox() function
This function first checks to see whether there is a worker instance corresponding to the given box number. If there is, it terminates the process running in that box, then clears the corresponding DOM element.

It is easier to clear the boxes in west.js instead of within the boxWorker object because WebWorker.terminate(), which can only be called in the main thread, stops a process immediately, which is the desired behavior. As far as I know, it's not possible to immediately halt a process from inside a WebWorker (self.close() waits until the task is finished.)