Fixed: unsubscribe completely
This deletes the topic key from the mei.Events cache instead of just removing the specific events subscribed to. This is useful in our case because we do not want to keep any previous subscriptions on a given topic when we hit Send -- we're only keeping the new subscriptions chosen by the user from the drop-down menus.